pyproject.toml
[project]
name = "robot-arm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "gpiozero>=2.0.1",
    "rpi-gpio>=0.7.1",
    "rpi-lgpio>=0.6",
]
